Added tests, moved to metpy/calc
Created tools.py, added test_tools.py in testing folder, fixed some
formatting issues as requested
