+++ metpy/calc/tests/test_tools.py
@@ -1,0 +1,16 @@
+# Copyright (c) 2008-2015 MetPy Developers.
+# Distributed under the terms of the BSD 3-Clause License.
+# SPDX-License-Identifier: BSD-3-Clause
+
+import numpy as np
+from numpy.testing import assert_array_equal
+
+from metpy.calc.tools import resample_nn_1d
+
+
+def test_resample_nn():
+    a = np.arange(5.)
+    b = np.array([2, 3.8])
+    truth = np.array([2, 4])
+
+    assert_array_equal(truth, resample_nn_1d(a, b))
