+++ metpy/calc/tools.py
@@ -1,0 +1,30 @@
+# Copyright (c) 2008-2015 MetPy Developers.
+# Distributed under the terms of the BSD 3-Clause License.
+# SPDX-License-Identifier: BSD-3-Clause
+
+import numpy as np
+
+
+def resample_nn_1d(a, centers):
+    """Helper function that returns one-dimensional nearest-neighbor
+    indexes based on user-specified centers.
+
+    Parameters
+    ----------
+    a : array-like
+        1-dimensional array of numeric values from which to
+        extract indexes of nearest-neighbors
+    centers : array-like
+        1-dimensional array of numeric values representing a subset of values to approximate
+
+    Returns
+    -------
+        An array of indexes representing values closest to given array values
+
+    """
+    ix = []
+    for center in centers:
+        index = (np.abs(a - center)).argmin()
+        if index not in ix:
+            ix.append(index)
+    return ix
