+++ metpy/plots/util.py
@@ -8,28 +8,6 @@ from matplotlib.collections import LineCollection
 from ..units import concatenate
 
 
-def resample_nn_1d(a, centers):
-    """Helper function that returns one-dimensional nearest-neighbor
-    indexes based on user-specified centers.
-    ----------
-    a : array-like
-        1-dimensional array of numeric values from which to
-        extract indexes of nearest-neighbors
-    centers : array-like
-        1-dimensional array of numeric values representing a subset of values to approximate
-    Returns
-    -------
-        An array of indexes representing values closest to given array values
-
-    """
-    ix = []
-    for center in centers:
-        index = (np.abs(a-center)).argmin()
-        if index not in ix:
-            ix.append(index)
-    return ix
-
-
 # Not part of public API
 def colored_line(x, y, c, **kwargs):
     """Helper function to take a set of points and turn them into a collection of
